- markup logic moved to a separate method
