+++ src/routes/router.ts
@@ -1,13 +1,11 @@
 import path from "path";
-import * as fs from "fs";
 import { Router, static as Static, Request, Response } from "express";
-import sanitizeHtml from "sanitize-html";
-import { marked } from "marked";
 import { connectAppDescriptor, connectDescriptorGet } from "./atlassian-connect";
 import { eventsRouter } from "./events";
 import { webhooksRouter } from "./webhooks";
 import { database } from "../db";
 import { connectIframeJWTMiddleware } from "../middlewares/connect-iframe-jwt-middleware";
+import getMarkdownAndConvertToHtml from "../utils/markup";
 
 export const RootRouter = Router();
 
@@ -25,31 +23,6 @@ RootRouter.use("/events", eventsRouter);
 
 // Jira webhooks we listen to as specified in the Connect JSON above
 RootRouter.use("/webhooks", webhooksRouter);
-
-const renderer = new marked.Renderer();
-
-renderer.link = (href: string, _, text: string): string => {
-	if (href?.includes("https" || href?.includes("http"))) {
-		return `<a target="_blank" href="${href}">${text}</a>`;
-	} else {
-		const page = href?.substring(1);
-		return `<span class="link-span" id="${page}" data-connect-module-key="${page}">${text}</span>`;
-	}
-};
-
-const getMarkdownAndConvertToHtml = (fileName: string): string => {
-	const filePath = path.join(__dirname, "..", "content", fileName);
-	const contents = fs.readFileSync(filePath);
-	// TODO - see if there's a way to modify the way we are using marked.js so we can pass data directly to HTML elements
-	const markdownToHtml = marked.parse(contents.toString(), { renderer: renderer });
-
-	return sanitizeHtml(markdownToHtml, {
-		allowedAttributes: {
-			span: [ "class", "id", "data-connect-module-key" ],
-			a: [ "href", "target" ]
-		}
-	});
-};
 
 RootRouter.use(connectIframeJWTMiddleware);
